Group Manager classes to packages. Rename calendarFetchStrategies package to strategies.
