+++ app/src/main/java/com/mymeetings/android/model/managers/CalendarEventAlertManager.kt
@@ -1,5 +1,9 @@
-package com.mymeetings.android.model
+package com.mymeetings.android.model.managers
 
+import com.mymeetings.android.model.CalendarEvent
+import com.mymeetings.android.model.CalendarEventPriority
+import com.mymeetings.android.model.CalendarEventWithViewAlert
+import com.mymeetings.android.model.ViewAlertType
 import com.mymeetings.android.utils.ClockUtils
 import java.util.concurrent.TimeUnit
 
@@ -30,6 +34,9 @@ class CalendarEventAlertManager(private val clockUtils: ClockUtils) {
             }
             else -> ViewAlertType.LOW
         }
-        return CalendarEventWithViewAlert(calendarEvent, viewAlert)
+        return CalendarEventWithViewAlert(
+            calendarEvent,
+            viewAlert
+        )
     }
 }
