+++ app/src/main/java/com/mymeetings/android/view/activities/ui/home/CalendarEventsViewModel.kt
@@ -1,11 +1,12 @@
 package com.mymeetings.android.view.activities.ui.home
 
 import androidx.lifecycle.ViewModel
-import com.mymeetings.android.model.CalendarEventsSyncManager
+import com.mymeetings.android.model.managers.CalendarEventsSyncManager
 import kotlinx.coroutines.*
 
 class CalendarEventsViewModel(
-    private val calendarEventsSyncManager: CalendarEventsSyncManager) : ViewModel() {
+    private val calendarEventsSyncManager: CalendarEventsSyncManager
+) : ViewModel() {
 
     private val backgroundScope = CoroutineScope(Dispatchers.IO)
